--- client/src/components/DocumentsComponent/components/Tabs/Contracts/index.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { MyContracts } from "../../../../ClientPage/components/Documents/MyContracts";
-import Box from '@material-ui/core/Box';
-
-export const Contracts = () => {
-
-  return (
-    <Box>
-      <MyContracts visibleFalse={true}/>
-    </Box>
-  );
-};
