+++ client/src/pages/Document/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { DocumentsComponent } from "../../components/DocumentsComponent";
+import { DocumentsComponent } from "../../components/DocumentsComponentUser";
 import  Container  from "../../components/Container";
 
 const Index =()=> {
